Backport following commits: 615: Envelope Generator outputs a short 616: Fix for Rayden/Boot_Zak_v2 617: Make clear method private, add some comments and remove unused include
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-0.3@620 75748be7-0387-4326-94c7-f41ce0883d2c

+++ builders/residfp-builder/residfp/EnvelopeGenerator.h
@@ -191,7 +191,7 @@ public:
 	 */
 	void clock();
 
-	int output() const;
+	short output() const;
 
 	/**
 	 * Constructor.
@@ -339,7 +339,7 @@ void EnvelopeGenerator::clock() {
 }
 
 RESID_INLINE
-int EnvelopeGenerator::output() const {
+short EnvelopeGenerator::output() const {
 	// DAC imperfections are emulated by using envelope_counter as an index
 	// into a DAC lookup table. readENV() uses envelope_counter directly.
 	return dac[envelope_counter & 0xff];
